requirements.txt
prometheus_client
psutil
python-dotenv
requests